+++ test/east-human/case27_class_instance.cpp
@@ -7,21 +7,21 @@ namespace east_view /* source: test/py/case27_class_instance.py */ {
 // [5:0] class original=Box100
 struct Box100 {
 // [6:4] function original=__init__
-None __init__(Box100 self /* mutable */, int seed /* readonly */) {
+None __init__(Box100 self /* mutable */, int64 seed /* readonly */) {
             // [7:8]
-            self.seed /* type=int, borrow=value */ = seed /* type=int, borrow=readonly_ref */;
+            self.seed /* type=int64, borrow=value */ = seed /* type=int64, borrow=readonly_ref */;
         }
 // [9:4] function original=next
-int next(Box100 self /* mutable */) {
+int64 next(Box100 self /* mutable */) {
             // [10:8]
-            self.seed /* type=int, borrow=value */ /* Add */= 1 /* type=int, borrow=value */;
+            self.seed /* type=int64, borrow=value */ /* Add */= 1 /* type=int64, borrow=value */;
             // [11:8]
-            return self.seed /* type=int, borrow=value */;
+            return self.seed /* type=int64, borrow=value */;
         }
     };
 
     // main guard body
-    int __east_main_guard() {
+    int64 __east_main_guard() {
         // [15:4]
         Box100 b /* type=Box100, borrow=value */ = Box100(3) /* type=Box100, borrow=value */;
         // [16:4]
